Title: 자연수 n개의 합, Time: 98ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 240804/자연수 n개의 합/sum-of-n-natural-numbers.kt
@@ -3,17 +3,17 @@ import kotlin.math.*
 
 fun main() {
     val br = BufferedReader(InputStreamReader(System.`in`))
-    val num = br.readLine().toInt()
-    var right = (sqrt(num.toDouble()).toInt() * 2)
-    var left = 0
-    var result =-1
-    while (left<right) {
+    val num = br.readLine().toLong()
+    var right = (sqrt(num.toDouble()).toLong() * 2)
+    var left = 0L
+    var result = -1L
+    while (left < right) {
         val mid = (right + left) / 2
         if ((mid * (mid - 1)) / 2 < num) {
-            left = mid+1
+            left = mid + 1
             result = mid
-        }else{
-            right = mid-1
+        } else {
+            right = mid - 1
         }
     }
     println(result)
